Fix bin_at() on Python 2
The bin_at() code was recently changed to use .format instead of the %
operator.  This requires an explicit int cast of m.address in order to
format it as hex, but the int cast only works on Python 3.  On Python 2
it raises an exception: "<class 'gdb.error'> Cannot convert value to
int".

The % operator (without an explicit cast) appears to work on Python 2
and Python 3 so use that.  Also add a comment explaining this so it
isn't inadvertently removed in the future.